RGB値に255 logのIntegerに Kpなどの値変更

+++ src/game/Log.java
@@ -125,11 +125,11 @@ public class Log {
             StringBuilder sb = new StringBuilder();
             // ヘッダー部
             sb.append("white,black,target\r\n");
-            sb.append(Float.toString(game.course.getRGB_White()));
+            sb.append(Integer.toString(game.course.getRGB_White()));
             sb.append(",");
-            sb.append(Float.toString(game.course.getRGB_Black()));
+            sb.append(Integer.toString(game.course.getRGB_Black()));
             sb.append(",");
-            sb.append(Float.toString(game.course.getRGB_Target()));
+            sb.append(Integer.toString(game.course.getRGB_Target()));
             sb.append("\r\n\r\n");
             // レコード部
             sb.append("time,status,RGB,forward,leftspeed,rightspeed\r\n");
@@ -138,7 +138,7 @@ public class Log {
                 sb.append(",");
                 sb.append(data.getStatus().toString());
                 sb.append(",");
-                sb.append(Float.toString(data.RGB_brightness()));
+                sb.append(Integer.toString(data.RGB_brightness()));
                 sb.append(",");
                 sb.append(Float.toString(data.getForward()));
                 sb.append(",");
